requirements.txt
langchain_google_genai
pypdf
pinecone-client
langchain
langchain_community
SpeechRecognition 
pydub 
google-cloud-speech